Renamed trackdata to inline_tiles
